Time: 68 ms (9.58%), Space: 85 MB (7.54%) - LeetHub

+++ 0940-fruit-into-baskets/0940-fruit-into-baskets.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     int totalFruit(vector<int>& fruits) {
-        long ans=0;
-        unordered_map<long,long> mmap={};
-        for (long left=0,right=0,tmp=0,l=0;right<fruits.size();right++){
+        unsigned int ans=0;
+        unordered_map<unsigned int,unsigned int> mmap={};
+        for (unsigned int left=0,right=0,tmp=0,l=0;right<fruits.size();right++){
             l++;
             if(mmap.contains(fruits[right])){
                 if(!mmap[fruits[right]]){
